+++ sdk/lib/convert/convert.dart
@@ -164,7 +164,8 @@ library dart.convert;
 
 import 'dart:async';
 import 'dart:typed_data';
-import 'dart:_internal' show CastConverter, checkNotNullable, parseHexByte;
+import 'dart:_internal'
+    show CastConverter, checkNotNullable, parseHexByte, Since;
 
 part 'ascii.dart';
 part 'base64.dart';